Recreate only when necessary

+++ packages/flutter_tools/lib/src/plugins.dart
@@ -898,9 +898,12 @@ Future<void> _writeWebPluginRegistrant(FlutterProject project, List<Plugin> plug
 /// For each platform that uses them, creates symlinks within the platform
 /// directory to each plugin used on that platform.
 ///
+/// If |force| is true, the symlinks will be recreated, otherwise they will
+/// be created only if missing.
+///
 /// This uses [project.flutterPluginsDependenciesFile], so should only be run
 /// after refreshPluginList has been run since the last plugin change.
-void createPluginSymlinks(FlutterProject project) {
+void createPluginSymlinks(FlutterProject project, {bool force = false}) {
   Map<String, dynamic> platformPlugins;
   final String pluginFileContent = _readFileContent(project.flutterPluginsDependenciesFile);
   if (pluginFileContent != null) {
@@ -912,20 +915,23 @@ void createPluginSymlinks(FlutterProject project) {
   if (featureFlags.isWindowsEnabled && project.windows.existsSync()) {
     _createPlatformPluginSymlinks(
       project.windows.pluginSymlinkDirectory,
-      platformPlugins[project.windows.pluginConfigKey] as List<dynamic>);
+      platformPlugins[project.windows.pluginConfigKey] as List<dynamic>,
+      force: force);
   }
   if (featureFlags.isLinuxEnabled && project.linux.existsSync()) {
     _createPlatformPluginSymlinks(
       project.linux.pluginSymlinkDirectory,
-      platformPlugins[project.linux.pluginConfigKey] as List<dynamic>);
+      platformPlugins[project.linux.pluginConfigKey] as List<dynamic>,
+      force: force);
   }
 }
 
-/// Replaces [symlinkDirectory] with a directory containing symlinks to each plugin
-/// listed in [platformPlugins].
-void _createPlatformPluginSymlinks(Directory symlinkDirectory, List<dynamic> platformPlugins) {
-  // Start fresh each time to avoid stale links.
-  if (symlinkDirectory.existsSync()) {
+/// Creates [symlinkDirectory] containing symlinks to each plugin listed in [platformPlugins].
+///
+/// If [force] is true, the directory will be created only if missing.
+void _createPlatformPluginSymlinks(Directory symlinkDirectory, List<dynamic> platformPlugins, {bool force = false}) {
+  if (force && symlinkDirectory.existsSync()) {
+    // Start fresh to avoid stale links.
     symlinkDirectory.deleteSync(recursive: true);
   }
   symlinkDirectory.createSync(recursive: true);
@@ -935,8 +941,12 @@ void _createPlatformPluginSymlinks(Directory symlinkDirectory, List<dynamic> pla
   for (final Map<String, dynamic> pluginInfo in platformPlugins.cast<Map<String, dynamic>>()) {
     final String name = pluginInfo[_kFlutterPluginsNameKey] as String;
     final String path = pluginInfo[_kFlutterPluginsPathKey] as String;
+    final Link link = symlinkDirectory.childLink(name);
+    if (link.existsSync()) {
+      continue;
+    }
     try {
-      symlinkDirectory.childLink(name).createSync(path);
+      link.createSync(path);
     } on FileSystemException catch (e) {
       if (globals.platform.isWindows && (e.osError?.errorCode ?? 0) == 1314) {
         throwToolExit(
@@ -965,7 +975,7 @@ void refreshPluginsList(FlutterProject project, {bool checkProjects = false}) {
 
   final bool changed = _writeFlutterPluginsList(project, plugins);
   if (changed || legacyChanged) {
-    createPluginSymlinks(project);
+    createPluginSymlinks(project, force: true);
     if (!checkProjects || project.ios.existsSync()) {
       cocoaPods.invalidatePodInstallOutput(project.ios);
     }
